% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{interpolate_iv_by_moneyness}
\alias{interpolate_iv_by_moneyness}
\title{PCHIP-Interpolation + Clamping (Python-kompatibel)}
\usage{
interpolate_iv_by_moneyness(mnes, vol, grid)
}
\description{
PCHIP-Interpolation + Clamping (Python-kompatibel)
}
\keyword{internal}
