% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_ref.R
\name{pyref_compute}
\alias{pyref_compute}
\title{DEV: compute via Python reference (not exported)}
\usage{
pyref_compute(mnes, vol, days, rate, params = rq_default_params())
}
\description{
DEV: compute via Python reference (not exported)
}
\keyword{internal}
