% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmoms.R
\name{qmoms_compute_bygroup}
\alias{qmoms_compute_bygroup}
\title{Compute by group (mirrors Python qmoms_compute_bygroup)}
\usage{
qmoms_compute_bygroup(
  groupparams,
  id = NULL,
  rate = NULL,
  days = NULL,
  date = NULL,
  cols_map = list(id = "id", date = "date", days = "days", rate = "rate", mnes = "mnes",
    impl_volatility = "impl_volatility")
)
}
\description{
Compute by group (mirrors Python qmoms_compute_bygroup)
}
