% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmoms_params.R
\name{rq_default_params}
\alias{rq_default_params}
\title{Default parameters (mirrors Python qmoms_params.default_params)}
\usage{
rq_default_params()
}
\description{
Default parameters (mirrors Python qmoms_params.default_params)
}
